jpa work effort

+++ src/main/java/com/microservices/app/h2/model/daos/H2ClientDao.java
@@ -1,6 +1,7 @@
 package com.microservices.app.h2.model.daos;
 
 import java.io.Serializable;
+
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -9,7 +10,6 @@ import javax.persistence.Id;
 import javax.persistence.Table;
 import javax.validation.constraints.NotNull;
 
-import lombok.AllArgsConstructor;
 import lombok.Data;
 
 @Data
@@ -21,7 +21,7 @@ public class H2ClientDao implements Serializable {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private long id;
+	private Integer id;
 	
 	@NotNull
 	@Column(name="NAME")
